+++ frontend/src/app/member/my/schedule/page.tsx
@@ -1,5 +1,10 @@
+import { Suspense } from "react";
 import ClientPage from "./ClientPage";
 
-export default async function Page() {
-  return <ClientPage />;
+export default function Page() {
+  return (
+    <Suspense fallback={<div>Loading...</div>}>
+      <ClientPage />
+    </Suspense>
+  );
 }
