+++ frontend/src/app/member/my/schedule/update/[id]/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-
+import { Suspense } from "react";
 import ClientPage from "./ClientPage";
 import Header from "@/components/Header";
 import Footer from "@/components/Footer";
@@ -8,7 +8,9 @@ export default function Page() {
   return (
     <>
       <Header />
-      <ClientPage />
+      <Suspense fallback={<div>로딩 중...</div>}>
+        <ClientPage />
+      </Suspense>
       <Footer />
     </>
   );
