+++ frontend/src/app/place/[id]/page.tsx
@@ -1,3 +1,4 @@
+import { Suspense } from "react";
 import Header from "@/components/Header";
 import Footer from "@/components/Footer";
 import ClientPage from "./ClientPage";
@@ -7,7 +8,9 @@ export default async function Page() {
     <div className="min-h-screen flex flex-col">
       <Header />
       <main className="flex-grow">
-        <ClientPage />
+        <Suspense fallback={<div>로딩 중...</div>}>
+          <ClientPage />
+        </Suspense>
       </main>
       <Footer />
     </div>
